Retter bug hvor lenke til dialog ikke fungerer i varsel og flytter litt kode (#703)
