+++ src/moduler/varslinger/VeilederVarsel.tsx
@@ -3,10 +3,8 @@ import { Normaltekst } from 'nav-frontend-typografi';
 import React from 'react';
 import { shallowEqual, useSelector } from 'react-redux';
 
-import { useReduxDispatch } from '../../felles-komponenter/hooks/useReduxDispatch';
-import InternLenke from '../../felles-komponenter/utils/InternLenke';
 import { formaterDatoKortManed } from '../../utils';
-import { velgHistoriskPeriode } from '../filtrering/filter/filter-reducer';
+import LenkeTilDialog from '../dialog/DialogLink';
 import {
     selectErEskalert,
     selectGjeldendeEskaleringsVarsel,
@@ -15,11 +13,9 @@ import {
 import styles from './varslinger.module.less';
 
 function VeilederVarsel() {
-    const dispatch = useReduxDispatch();
     const erEskalert = useSelector(selectErEskalert);
     const dialogId = useSelector(selectTilHorendeDialogId);
     const eskaleringsVarsel = useSelector(selectGjeldendeEskaleringsVarsel, shallowEqual);
-    const doVelgNavarendePeriode = () => dispatch(velgHistoriskPeriode(null));
 
     if (!erEskalert || !eskaleringsVarsel) {
         return null;
@@ -30,10 +26,7 @@ function VeilederVarsel() {
         <div className="container">
             <AlertStripeAdvarsel className={styles.varsling}>
                 <Normaltekst>
-                    NAV har sendt varsel {dato}{' '}
-                    <InternLenke href={`/dialog/${dialogId}`} onClick={doVelgNavarendePeriode}>
-                        Les meldingen
-                    </InternLenke>
+                    NAV har sendt varsel {dato} <LenkeTilDialog dialogId={dialogId}>Les meldingen</LenkeTilDialog>
                 </Normaltekst>
             </AlertStripeAdvarsel>
         </div>
